refactor: move diffLineText into command_line_text module
